chore: remove container folder, delete unused files and variables, add more tests
