+++ src/Altinn.Apps/AppFrontend/react/receipt/src/index.tsx
@@ -1,8 +1,6 @@
-import * as React from 'react';
+import React from 'react';
 import { render } from 'react-dom';
-import App from './App';
 
-render(
-  <App />,
-  document.getElementById('root'),
-);
+import { App } from './App';
+
+render(<App />, document.getElementById('root'));
