+++ src/Altinn.Apps/AppFrontend/react/receipt/src/utils/instance.test.tsx
@@ -1,13 +1,45 @@
-import 'jest';
-import { getInstanceId, getInstanceOwnerId } from './instance';
+import { getInstanceId, getInstanceOwnerId, getArchiveRef } from './instance';
 
-describe('>>> utils/instance.test.tsx', () => {
-  // instanceId and instanceOwnerId are set in package.json under "jest"
-  it('should return correct instanceOwnerId', () => {
-    expect(getInstanceId()).toEqual('6697de17-18c7-4fb9-a428-d6a414a797ae');
+import { mockLocation } from 'testConfig/testUtils';
+
+const originalLocation = window.location;
+
+// instanceId and instanceOwnerId are set in package.json under "jest.testURL"
+describe('utils/instance', () => {
+  beforeEach(() => {
+    mockLocation(originalLocation);
+  });
+  describe('getInstanceId', () => {
+    it('should return instanceOwnerId when it exists in url', () => {
+      expect(getInstanceId()).toEqual('6697de17-18c7-4fb9-a428-d6a414a797ae');
+    });
+
+    it('should return empty string when window.location.pathname is not set', () => {
+      mockLocation({ pathname: undefined });
+      expect(getInstanceId()).toEqual('');
+    });
   });
 
-  it('should return correct instanceId', () => {
-    expect(getInstanceOwnerId()).toEqual('mockInstanceOwnerId');
+  describe('getInstanceOwnerId', () => {
+    it('should return instanceOwnerId when it exists in url', () => {
+      expect(getInstanceOwnerId()).toEqual('mockInstanceOwnerId');
+    });
+
+    it('should return empty string when window.location.pathname is not set', () => {
+      mockLocation({ pathname: undefined });
+      expect(getInstanceOwnerId()).toEqual('');
+    });
+  });
+
+  describe('getArchiveRef', () => {
+    it('should return last part of instanceId', () => {
+      expect(getArchiveRef()).toEqual('d6a414a797ae');
+    });
+
+    it('should return undefined when no instanceId is set', () => {
+      mockLocation({ pathname: undefined });
+
+      expect(getArchiveRef()).toEqual(undefined);
+    });
   });
 });
